dom/media/tests/mochitest/TextTrack_removeCue.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
  <title>Test for Bug 833386</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="start()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=833386">Mozilla Bug 833386</a>
<p id="display"></p>
<div id="content" style="display: none">
  <video src="demo.webm">
    <track id="track_1" kind="subtitles" src="fast.vtt" default />
  </video>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
  SpecialPowers.setBoolPref("media.webvtt.enabled", true);
  SimpleTest.waitForExplicitFinish();
  
  var textTrack;

  function start() {
    var trackElement = document.getElementById('track_1');
    textTrack = trackElement.track;

    is(textTrack.cues.length, 3);

    // remove invalid cue
    try {
      cue = new TextTrackCue(1, 2, 'foo');
      textTrack.removeCue(cue);
    } catch (e) {
      is(e.name, "NotFoundError");
    }

    // remove valid cue
    cue = textTrack.cues[0];
    textTrack.removeCue(cue);
    is(textTrack.cues.length, 2);

    SimpleTest.finish();
  }
  addLoadEvent(start);
</script>
</pre>
</body>
</html>
